Kill stupid code
