+++ mobile/src/main/rust/vpnhotspotd/src/lib.rs
@@ -12,6 +12,3 @@ mod ra;
 #[cfg(test)]
 #[allow(dead_code)]
 mod socket;
-#[cfg(test)]
-#[allow(dead_code)]
-mod upstream;
